iterator refactor
